backend/requirements.txt
fastapi
uvicorn[standard]
praw
tweepy
yfinance
httpx
python-dotenv
anthropic
beautifulsoup4
